pyproject.toml
[tool.poetry]
name = "peckis"
version = "0.1.0"
description = "Peckis Doodler Backend"
authors = ["Matas Seimys <matas.seimys@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.6"
flask = "^1.0"
flask-cors = "^3.0"
Pillow = "^5.4"
pytest = "^4.3"
pytest-cov = "^2.6"
pytest-flask = "^0.14.0"
opencv-python = "^4.1"
celery = "^4.3"
sqlalchemy = "^1.3"

[tool.poetry.dev-dependencies]
black = {version = "^18.3-alpha.0",allows-prereleases = true}

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
